Use size_t consistently, as we dropped fmin()

+++ src/buffer.c
@@ -6,7 +6,7 @@ void mqtt_buffer_dump(mqtt_buffer_t* buffer) {
   printf("[%d] ", buffer->length);
 
   char hex = 0;
-  for (int i=0;i<buffer->length;++i) {
+  for (size_t i=0;i<buffer->length;++i) {
     if (buffer->data[i] < 0x20 || buffer->data[i] > 0x7e) {
       hex = 1;
       break;
@@ -21,13 +21,13 @@ void mqtt_buffer_dump(mqtt_buffer_t* buffer) {
 }
 
 void mqtt_buffer_dump_ascii(mqtt_buffer_t* buffer) {
-  for (int i=0;i<buffer->length;++i) {
+  for (size_t i=0;i<buffer->length;++i) {
     printf("%c", buffer->data[i]);
   }
 }
 
 void mqtt_buffer_dump_hex(mqtt_buffer_t* buffer) {
-  for (int i=0;i<buffer->length;++i) {
+  for (size_t i=0;i<buffer->length;++i) {
     printf("%02x ", buffer->data[i]);
   }
 }
